Move `PoolTileMapper` to GPU.
The process of mapping the average color of a tile on the image, to an image from the pool was taking a long time.
This operation was previously done serially, by looping through all of the tiles, and for each tile looping through all of the images from the pool to find the best match.
This is now fully done on the GPU in parallel, which is much faster.