Fix GPU stitching.
Multiple things were not going right with stitching.
We were for some odd reason drawing the pictures only of the size of the number of tiles.
This makes no sense. We should obviously draw the size of the full tile image.
We also don't pass in the position of the image anymore, since we have it through calculations.