Fix bug where black borders would be added to drawn tiles.
We were trying to fit a certain amount of tiles into a fix size.
For example, we would try to fit 3 tiles into 10 pixels, which is impossible.
To compensate, we would add an extra black pixel.
This resulted in black border being drawn around some tiles.

To fix this, we now make sure to output an image that fits exactly the number of tiles.
To continue with the above example, we now generate a 9x9 image instead of a 10x10 one.
This ensures that all 3 tiles fit perfectly.